FEAT(history): change history and capture history types from int32_t to int16_t for reduced memory footprint and improved performance (used  encodedMove from TT to get reduced size)

+++ engine/search/search_entry.cpp
@@ -586,7 +586,7 @@ chess::Board::Move Engine::searchUCI(uint64_t requestedDepth) noexcept {
     this->nodesSearched = 0;
 
     // History soft reset (same as search())
-    int32_t* historyFlat = &this->history[0][0][0];
+    int16_t* historyFlat = &this->history[0][0][0];
     constexpr int HISTORY_CELLS = 2 * 64 * 64;
     #pragma omp simd
     for (int i = 0; i < HISTORY_CELLS; ++i) {
@@ -640,7 +640,7 @@ void Engine::search(uint64_t requestedDepth) noexcept {
     // Prevent stale data from dominating move ordering
     // Divide all history values by 2 at the start of each new search
     // This gives recent data more weight while preserving good moves
-    int32_t* historyFlat = &this->history[0][0][0];
+    int16_t* historyFlat = &this->history[0][0][0];
     constexpr int HISTORY_CELLS = 2 * 64 * 64;
     #pragma omp simd
     for (int i = 0; i < HISTORY_CELLS; ++i) {
